requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.12.0
openai>=1.52.0
ollama>=0.3.3
numpy>=2.4.0
pytest>=8.3.3
pyyaml>=6.0.3
httpx>=0.27.2
python-dotenv>=1.0.1
openenv-core